Break items out into seperate components.
