+++ src/components/canopy/MetadataDropdown.js
@@ -2,12 +2,9 @@ import React, { Component } from 'react';
 import * as DropdownMenu from "@radix-ui/react-dropdown-menu"
 import { BiChevronDown } from "react-icons/bi"
 import { StaticQuery, graphql } from "gatsby"
-import { countBy } from 'lodash';
+import MetadataDropdownContent from "./MetadataDropdownContent"
 
 class MetadataDropdown extends Component {
-  constructor(props) {
-    super(props);
-  }
 
   renderContent(label) {
     return (
@@ -58,33 +55,6 @@ class MetadataDropdown extends Component {
       </div>
     )
   }
-}
-
-const MetadataDropdownContent = ({ data, label }) => {
-  let values = []
-  const counts = {}
-  for (const result of data.allManifests.edges) {
-    for (const elemMatch of result.node.metadata) {
-      if (elemMatch.label.en[0] === label) {
-        for (const string of elemMatch.value.en) {
-          values.push(string)
-        }
-      }
-    }
-  }
-  values.sort();
-  const content = countBy(values)
-  return Object.keys(content).map((value) => {
-    return <MetadataDropdownContentItem value={value} count={content[value]}/>
-  });
-}
-
-const MetadataDropdownContentItem = ({ value, count }) => {
-  return (
-    <a>
-      <span>{count}</span> {value}
-    </a>
-  )
 }
 
 export default MetadataDropdown;
